cleanup index - check db version externally define required db version in cofig as constant some type updates
