+++ backend/src/server/createServer.ts
@@ -1,15 +1,12 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-
 import 'reflect-metadata'
 import 'module-alias/register'
 
 import { ApolloServer } from 'apollo-server-express'
-import express from 'express'
+import express, { Express } from 'express'
 
 // database
 import connection from '../typeorm/connection'
-import getDBVersion from '../typeorm/getDBVersion'
+import { checkDBVersion } from '../typeorm/DBVersion'
 
 // server
 import cors from './cors'
@@ -24,27 +21,25 @@ import schema from '../graphql/schema'
 
 // webhooks
 import { elopageWebhook } from '../webhook/elopage'
+import { Connection } from 'typeorm'
 
 // TODO implement
 // import queryComplexity, { simpleEstimator, fieldConfigEstimator } from "graphql-query-complexity";
 
-const DB_VERSION = '0012-login_user_backups_unify_wordlist'
+type ServerDef = { apollo: ApolloServer; app: Express; con: Connection }
 
-const createServer = async (context: any = serverContext): Promise<any> => {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const createServer = async (context: any = serverContext): Promise<ServerDef> => {
   // open mysql connection
   const con = await connection()
   if (!con || !con.isConnected) {
-    throw new Error(`Couldn't open connection to database`)
+    throw new Error(`Fatal: Couldn't open connection to database`)
   }
 
   // check for correct database version
-  const dbVersion = await getDBVersion()
-  if (!dbVersion || dbVersion.indexOf(DB_VERSION) === -1) {
-    throw new Error(
-      `Wrong database version - the backend requires '${DB_VERSION}' but found '${
-        dbVersion || 'None'
-      }'`,
-    )
+  const dbVersion = await checkDBVersion(CONFIG.DB_VERSION)
+  if (!dbVersion) {
+    throw new Error('Fatal: Database Version incorrect')
   }
 
   // Express Server
@@ -57,15 +52,6 @@ const createServer = async (context: any = serverContext): Promise<any> => {
   app.use(express.json())
   // bodyparser urlencoded for elopage
   app.use(express.urlencoded({ extended: true }))
-
-  // Log every request
-  /*
-  app.use((req, res, next) => {
-    // eslint-disable-next-line no-console
-    console.log(req)
-    next()
-  })
-  */
 
   // Elopage Webhook
   app.post('/hook/elopage/' + CONFIG.WEBHOOK_ELOPAGE_SECRET, elopageWebhook)
